Updater: Cleanup progress based on manifest stream
