+++ lib/update_util/resources/manifest_i.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <toolbox/stream/buffered_file_stream.h>
+
+struct ResourceManifestReader {
+    Storage* storage;
+    Stream* stream;
+    FuriString* linebuf;
+    ResourceManifestEntry entry;
+};
